.privateclaw/.scripts/pyproject.toml
[project]
name = "privateclaw"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "pyannote.audio",
    "pytesseract",
    "pymupdf",
    "Pillow",
    "ollama",
    "torchaudio",
    "python-dotenv",
]

[project.scripts]
pc-transcribe = "privateclaw.transcribe:main"
pc-flag = "privateclaw.flag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
constraint-dependencies = [
    "torch>=2.0",
    "torchaudio>=2.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
